Move json data
Move loaders for parameter defaults and descriptions to
dedicated grblogtools.parameters package
