+++ setup.cfg
@@ -37,4 +37,4 @@ include_package_data = True
 where = src
 
 [options.package_data]
-grblogtools.defaults = *.json
+grblogtools.parameters.data = *.json
