+++ src/grblogtools/parameters/__init__.py
@@ -1,0 +1,2 @@
+from grblogtools.parameters.defaults import load_defaults
+from grblogtools.parameters.pretty import load_descriptions
